refactor index.js, create src folder for core modules
