+++ lib/client.js
@@ -1,12 +1,12 @@
 // The client for js; imports all js modules
-
-// dependencies
 const Promise = require('bluebird')
 const _ = require('lomath')
 const path = require('path')
-const { setEnv, log } = require(path.join(__dirname, '..', 'index'))
-
+const SRCPATH = path.join(__dirname, '..', 'src')
+const log = require(path.join(SRCPATH, 'log'))
+const { setEnv } = require(path.join(SRCPATH, 'env'))
 global.hasher = require(path.join(__dirname, 'js', 'hasher')) // the js handler hasher
+
 
 /* istanbul ignore next */
 if (process.env.IOPORT == undefined) { setEnv() }
@@ -35,9 +35,7 @@ function correctReply(reply, msg) {
 }
 
 
-//////////////////////////////////////
-// 1. Register the socket.io client //
-//////////////////////////////////////
+// 1. Register the socket.io client
 /* istanbul ignore next */
 var IOPORT = process.env.IOPORT || 6466
 const client = require('socket.io-client')('http://localhost:' + IOPORT);
@@ -48,16 +46,11 @@ client.emit('join', ioid)
 client.on('disconnect', client.disconnect)
 
 
-//////////////////////////////////////////////////////
-// 2. Write module methods and register as handlers //
-//////////////////////////////////////////////////////
+// 2. Write module methods and register as handlers
 // done in your module scripts
 
 
-/////////////////////////////////////////////
-// 3. listener to handle incoming payload. //
-/////////////////////////////////////////////
-
+// 3. listener to handle incoming payload.
 /**
  * The handle for client.on('take') on getting replies
  * @param  {JSON} msg Replied msg
@@ -95,9 +88,7 @@ function handle(msg) {
 client.on('take', handle)
 
 
-////////////////////////////////////////
-// node.js entry-point client methods //
-////////////////////////////////////////
+// node.js entry-point client methods
 
 /**
  * Pass that returns a promise for getting client replies conveniently.
